Added Redstone Circuits and a Farming System to Expand Game Features and Interaction - Added script references related to Redstone circuits and the farming system in HTML to support the loading of new features. - Implemented basic Redstone circuit logic in JavaScript, including interactions between components such as wires, switches, and gates. - Implemented the farming system, supporting crop planting, growth, and harvesting to enhance the game’s survival experience. - Updated the README document to provide detailed descriptions of how to use the new features and their interaction logic. - Optimized the relevant script versions to ensure compatibility and stability of the new features.

+++ js/world/infinite.js
@@ -79,7 +79,7 @@ window.Voxel = window.Voxel || {};
   function newChunk(cx, cz) {
     return {
       cx: cx, cz: cz,
-      blocks: new Uint8Array(CS * H * CS),
+      blocks: new Uint16Array(CS * H * CS), // v8：方块 ID 空间 65535（与 world.js/gen_core 一致）
       heights: new Int16Array(CS * CS),
       biomes: new Uint8Array(CS * CS),
       sky: new Uint8Array(CS * H * CS),
@@ -303,7 +303,7 @@ window.Voxel = window.Voxel || {};
       if (coreBlkOverlay) { coreBlkOverlay = null; markCoreEdgeMeshes(); }
       return;
     }
-    if (!coreBlkOverlay) coreBlkOverlay = new Uint8Array(W * H * D);
+    if (!coreBlkOverlay) coreBlkOverlay = new Uint16Array(W * H * D);
     else coreBlkOverlay.fill(0);
     var buckets = [];
     for (var cz = 0; cz < CORE_CZ; cz++) {
@@ -541,7 +541,7 @@ window.Voxel = window.Voxel || {};
     var k = key(cx, cz);
     var ch = extra[k];
     if (!ch) { ch = newChunk(cx, cz); extra[k] = ch; }
-    ch.blocks = new Uint8Array(blocksBuf);
+    ch.blocks = new Uint16Array(blocksBuf);
     ch.heights = new Int16Array(heightsBuf);
     ch.biomes = new Uint8Array(biomesBuf);
     ch.baseReady = true;
